api/requirements.txt
fastapi
uvicorn
pillow
httpx
numpy
